Corrected types and modified training algorithm

+++ src/main/scala/net/ddns/akgunter/rl4j_practice/environment/BanditEnv.scala
@@ -21,7 +21,7 @@ class BanditEnv(numBandits: Int, numMachines: Int, maxNumSteps: Int, debug: Bool
 
   override def getActionSpace: DiscreteSpace = actionSpace
 
-  def getRewards: Array[Array[Int]] = currentState.getRewards
+  def getRewards: Array[Array[Double]] = currentState.getRewards
 
   def getBestMachines: Array[Int] = currentState.getBestMachines
